+++ packages/tools/CHANGELOG.md
@@ -1,5 +1,15 @@
 # librecode-tools
 
+## 0.3.5
+
+### Patch Changes
+
+- fix: add /setup command to TUI and restore direct arguments to /provider command
+
+- Updated dependencies []:
+  - librecode-types@0.3.5
+  - librecode-utils@0.3.5
+
 ## 0.3.4
 
 ### Patch Changes